third_party/blink/web_tests/platform/mac-mac11-arm64/virtual/background-svg-in-lcp/external/wpt/largest-contentful-paint/image-src-change-expected.txt

This is a testharness.js-based test.
FAIL Largest Contentful Paint: changing src causes a new entry to be dispatched. assert_between_exclusive: loadTime should occur between the lower bound and the renderTime expected a number greater than 21.200000002980232 and less than 21.799 but got 21.2
Harness: the test ran to completion.
